nicehash_btc_bal_function/requirements.txt
myb-nicehash-api>=1.0.0
# For getting BTC prices in Fiat from Coindesk
myb-coindesk-api>=1.0.0
myb-newrelic-api>=1.0.0
